requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
--trusted-host pypi.ngc.nvidia.com

aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
asciitree==0.3.3
    # via zarr
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
fasteners==0.19
    # via zarr
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.10.0
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   torch
huggingface-hub==0.26.1
    # via
    #   -r requirements.in
    #   timm
    #   tokenizers
    #   transformers
hydra-core==1.3.2
    # via -r requirements.in
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.4
    # via torch
lightning-utilities==0.11.8
    # via
    #   pytorch-lightning
    #   torchmetrics
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numcodecs==0.13.1
    # via zarr
numpy==1.26.4
    # via
    #   numcodecs
    #   torchmetrics
    #   torchvision
    #   transformers
    #   xformers
    #   zarr
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
omegaconf==2.3.0
    # via hydra-core
packaging==24.1
    # via
    #   huggingface-hub
    #   hydra-core
    #   lightning-utilities
    #   pytorch-lightning
    #   torchmetrics
    #   transformers
pillow==11.0.0
    # via torchvision
propcache==0.2.0
    # via yarl
pytorch-lightning==2.4.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   omegaconf
    #   pytorch-lightning
    #   timm
    #   transformers
regex==2024.9.11
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.4.5
    # via
    #   timm
    #   transformers
sympy==1.13.1
    # via torch
timm==1.0.11
    # via -r requirements.in
tokenizers==0.20.1
    # via transformers
torch==2.5.0
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   timm
    #   torchmetrics
    #   torchvision
    #   xformers
torchmetrics==1.5.0
    # via
    #   -r requirements.in
    #   pytorch-lightning
torchvision==0.20.0
    # via
    #   -r requirements.in
    #   timm
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
transformers==4.45.2
    # via -r requirements.in
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   multidict
    #   pytorch-lightning
    #   torch
urllib3==2.2.3
    # via requests
xformers==0.0.28.post2
    # via -r requirements.in
yarl==1.16.0
    # via aiohttp
zarr==2.18.3
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
